Extract selection methods from text_area/mod.rs into selection.rs
Move has_selection, selection_positions, selected_text, clipboard,
clear_selection, ensure_selection_anchor, and delete_selection into
a dedicated selection.rs submodule. This follows the same pattern
used by cursor.rs and brings mod.rs from 1012 to 939 lines,
satisfying the project's 1000-line limit.
